#  [JIT-1711] Compare the unprocessed transaction storage BundleStorage against a constant instead of VecDeque::capacity()