--- a/core/src/bundle_stage/bundle_consumer.rs
+++ b/core/src/bundle_stage/bundle_consumer.rs
@@ -835,7 +835,7 @@ mod tests {
         },
         solana_streamer::socket::SocketAddrSpace,
         std::{
-            collections::{HashSet, VecDeque},
+            collections::HashSet,
             str::FromStr,
             sync::{
                 atomic::{AtomicBool, Ordering},
@@ -1089,10 +1089,7 @@ mod tests {
 
         let bank_start = poh_recorder.read().unwrap().bank_start().unwrap();
 
-        let mut bundle_storage = UnprocessedTransactionStorage::new_bundle_storage(
-            VecDeque::with_capacity(10),
-            VecDeque::with_capacity(10),
-        );
+        let mut bundle_storage = UnprocessedTransactionStorage::new_bundle_storage();
         let mut bundle_stage_leader_metrics = BundleStageLeaderMetrics::new(1);
 
         let mut packet_bundles = make_random_overlapping_bundles(
@@ -1242,10 +1239,7 @@ mod tests {
 
         let bank_start = poh_recorder.read().unwrap().bank_start().unwrap();
 
-        let mut bundle_storage = UnprocessedTransactionStorage::new_bundle_storage(
-            VecDeque::with_capacity(10),
-            VecDeque::with_capacity(10),
-        );
+        let mut bundle_storage = UnprocessedTransactionStorage::new_bundle_storage();
         let mut bundle_stage_leader_metrics = BundleStageLeaderMetrics::new(1);
         // MAIN LOGIC